move sql query to db
